Minor improvement: Changed path of the service config file for Debian
On Debian/Ubuntu systems, the suggested config path for the service
configuration is /etc/supervisor/conf.d.
Changed the path for Debian system to this location. Also changed the
file extension from .ini to .conf.
The reason of this renaming is the naming inconsistency of the main
supervisor.conf file and the per service conf file (.ini).