Fixed: 'Could not restart Service' error.
Without this fix, attempting to restart a supervisor service on Ubuntu 12.04 would result in a "Could not restart Service" error, even that the restart was in fact successful. The \n in the grep pattern was being actualised by puppet, rather than being made a part of the regex.

This fix makes the \n part of the stopped … started regex.